feat: loadable pallet with empty execute extrinsic
This commit introduces truly working pallet that can be loaded into
the Substrate runtime.

Pallet can be loaded and execute function can be run which is emitting
event about the fact it has been called.

Weight file has been generated automatically as weights are needed
for the extrinsics to work.