test: unit tests for rpc crate added
There is a little chance to add more tests as most functions demand
runtime to work. There is no point to create a runtime working for
unit tests as we'll do it for the integration tests.
Each function demanding runtime has to be tested during integration
testing as we need to know how well it will cooperate with the runtime
and Substrate framework itself.
Therefore, unit tests will be added only for those functions which can
be executed without creating runtime environment.